# Close response body correctly and optimize base58 decode