### Problem

1. Request body was not being closed properly for non-200 responses.
2. Unneeded map usage and allocations in `base58` package.
3. `client.Ping` does not close tcp connections (no fix) -- should be addressed. Will cause too many open files and possible memory leak.

### Solution

1. Move `defer response.Body.Close()` before status code. Better tests should be added, but this fixes a memory leak.
2. Couple of changes here:
- convert `defaultAlphabet` to a hash table -- safe to be read concurrently, no need to copy map for each key.
- use only slice of bytes instead of `bytes.Buffer` for return payload
3. No change. A solution should be planned

Before:
```
go test -bench=. -benchmem -cpu 1,2,4
BenchmarkBase58           100000             20614 ns/op            6128 B/op         86 allocs/op
BenchmarkBase58-2         100000             19076 ns/op            6128 B/op         86 allocs/op
BenchmarkBase58-4         100000             19707 ns/op            6127 B/op         86 allocs/op
```

After:
```
go test -bench=. -benchmem -cpu 1,2,4
BenchmarkBase58           200000              7300 ns/op            2784 B/op         70 allocs/op
BenchmarkBase58-2         200000              6725 ns/op            2784 B/op         70 allocs/op
BenchmarkBase58-4         200000              6834 ns/op            2784 B/op         70 allocs/op
```

### Remember

- Bump `VERSION`
- Update `CHANGELOG.md`
- Update `README.md`
- Write tests