@revett 2017-09-11
Thanks for the PR @eramus! Checking it out now, sorry about the delay.