@revett 2017-09-11
@eramus I added a hotfix for the `.Ping()` close issue with #9. Thanks for spotting it!